requirements.txt
beautifulsoup4==4.11.1
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
streamlit==1.26.0
